+++ AuRoRA/src/vcs/CHANGELOG.rst
@@ -10,6 +10,12 @@ Forthcoming
 -----------
 * Added launch file to launch vtc and vcc nodes based on role parameter
 * Changed the threaded terminal print to use rclpy logger
+* Added QoS Profile to vital pulse signal and response
+* Implemented PRO2 architectural for VTC
+* - Pump: Collects raw sensor data for different sensors using Tripe-Level Polling (TLP) technique
+* - Regulator: Normalizes raw sensor data into proper format and regulates the oscillation rhythm according to RTT
+* - Oscillator: Packs and encodes sensor data into vital pulse signal and publishes at oscillation rhythm
+* - Orchestrator: Monitors vital pulse signal and response, detects disconnections, and triggers safety interlocks
 
 0.1.2 (2026-01-01)
 ------------------